+++ types/component.d.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 
 export class WeuiComponent extends Vue {
-    static name: string
+  static name: string
 
-    static install(vue: typeof Vue): void
+  static install(vue: typeof Vue): void
 }